+++ src/test/java/bridge/BridgeMakerTest.java
@@ -1,5 +1,6 @@
 package bridge;
 
+import bridge.domain.BridgeMaker;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -20,13 +21,13 @@ class BridgeMakerTest {
     @Test
     void 다리만들기_테스트() {
         List<String> bridge = bridgeMaker.makeBridge(3);
-        assertEquals(3,bridge.size());
+        assertEquals(3, bridge.size());
     }
 
     @Test
     void 숫자_입력시_문자로변환_테스트() {
-        assertEquals("U",bridgeMaker.changeNumber(1));
-        assertEquals("D",bridgeMaker.changeNumber(0));
+        assertEquals("U", bridgeMaker.changeNumber(1));
+        assertEquals("D", bridgeMaker.changeNumber(0));
     }
 
     public class TestNumberGenerator implements BridgeNumberGenerator {
